requirements.txt
fastapi>=0.111.0
python-multipart>=0.0.9
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
langgraph>=0.1.0
langchain-core>=0.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0
docker>=7.1.0
python-dotenv>=1.0.0
litellm>=1.40.0
httpx>=0.27.0
psutil>=5.9.0
